Add account modal, view all branches method & eventlisteners

+++ src/tree.js
@@ -7,6 +7,9 @@ const treeList = document.createElement('div');
 const branchList = document.createElement('div');
 const branchWrapper = document.createElement('div');
 const treeWrapper = document.createElement('div');
+let treeHeader = document.createElement('h2');
+let branchHeader = document.createElement('h2');
+let viewAll = document.createElement('button');
 
 const formTree = new PForm();
 const formBranch = new BForm();
@@ -16,6 +19,39 @@ const updateBranch = new BForm();
 branchWrapper.className = 'branch-wrap';
 branchList.className = 'branch-list';
 treeWrapper.className = 'tree-list';
+treeHeader.textContent = 'Your Projects';
+branchHeader.textContent = 'Project Todos';
+viewAll.type = 'button';
+viewAll.classList.add('btn', 'primary');
+viewAll.textContent = 'View All Todos';
+
+treeWrapper.appendChild(treeHeader);
+treeWrapper.appendChild(viewAll);
+branchWrapper.appendChild(branchList);
+
+function addAll(user){
+  branchList.innerHTML= '';
+  let allHeader = document.createElement('h1');
+  allHeader.textContent = 'All Todos';
+  branchList.appendChild(allHeader);
+  const allBranches = (() => {
+    db.collection('branches').where('user_id', '==', user.uid).orderBy('due_date').onSnapshot(bsnapshot => {
+      let bchanges = bsnapshot.docChanges();
+      bchanges.forEach(bchange => {
+        if(bchange.type == 'added'){
+          renderBranches(bchange.doc);
+        }else if(bchange.type == 'modified'){
+          renderBranches(bchange.doc);
+        }else if(bchange.type == 'removed'){
+          let li = branchList.querySelector('[data-id=' + bchange.doc.id + ']');
+          branchList.removeChild(li);
+        }
+      })
+    }, err => {
+      console.log(err.message)
+    })
+  })();
+} 
 
 function treeUpdate(el, id){
   el.appendChild(updateTree.content)
@@ -46,45 +82,51 @@ function branchUpdate(el, id){
 }
 
 function renderBranches(branch){
-  const newBranch = new Branch(
-    branch.data().priority,
-    branch.data().notes,
-    branch.data().title,
-    branch.data().description,
-    branch.data().due_date
-  )
-
-  newBranch.content.setAttribute('data-id', branch.id);
-
-  newBranch.showmore.addEventListener('click', (e) => {
-    e.stopPropagation();
-    showDes(newBranch.des);
-  })
-
-  newBranch.remove.addEventListener('click', (e) => {
-    e.stopPropagation();
-    let id = e.target.parentElement.getAttribute('data-id');
-    db.collection('branches').doc(id).delete();
-  })
-  newBranch.update.addEventListener('click', (e) => {
-    e.stopPropagation();
-    let id = e.target.parentElement.getAttribute('data-id');
-    branchUpdate(li, id);
-  })
-  newBranch.status.addEventListener('change', () => {
-    if(this.checked){
-      newBranch.content.style.background = 'lightgreen';
-      db.collection('branches').doc(branch.id).update({
-          status: true
-      });
-    }else{
-      newBranch.content.style.background = 'lightcoral';
-      db.collection('branches').doc(branch.id).update({
-        status: false
-      });
-    }
-  })
-  branchList.appendChild(newBranch.content);
+  if(branch){
+    const newBranch = new Branch(
+      branch.data().priority,
+      branch.data().notes,
+      branch.data().title,
+      branch.data().description,
+      branch.data().due_date
+    )
+  
+    newBranch.content.setAttribute('data-id', branch.id);
+  
+    newBranch.showmore.addEventListener('click', (e) => {
+      e.stopPropagation();
+      showDes(newBranch.des);
+    })
+  
+    newBranch.remove.addEventListener('click', (e) => {
+      e.stopPropagation();
+      let id = e.target.parentElement.getAttribute('data-id');
+      db.collection('branches').doc(id).delete();
+    })
+    newBranch.update.addEventListener('click', (e) => {
+      e.stopPropagation();
+      let id = e.target.parentElement.getAttribute('data-id');
+      branchUpdate(li, id);
+    })
+    newBranch.status.addEventListener('change', () => {
+      if(this.checked){
+        newBranch.content.style.background = 'lightgreen';
+        db.collection('branches').doc(branch.id).update({
+            status: true
+        });
+      }else{
+        newBranch.content.style.background = 'lightcoral';
+        db.collection('branches').doc(branch.id).update({
+          status: false
+        });
+      }
+    })
+    branchList.appendChild(newBranch.content);
+  }else{
+    let branchMessage = document.createElement('h2');
+    branchMessage.textContent = 'No Todos, create one from projct list!'
+    branchList.appendChild(branchMessage);
+  }
 }
 
 function showBranches(parent, id){
@@ -104,36 +146,43 @@ function showBranches(parent, id){
   }, err => {
     console.log(err.message)
   })
+  branchList.appendChild(branchHeader);
   parent.appendChild(formBranch.content);
-  branchWrapper.appendChild(branchList);
+  //branchWrapper.appendChild(branchList);
 }
 
 function renderTree(doc){
-  const newTree = new Tree(
-    doc.data().title,
-    doc.data().description,
-    doc.data().color
-  )
-
-  newTree.content.setAttribute('data-id', doc.id);
-
-  newTree.remove.addEventListener('click', (e) => {
-    e.stopPropagation();
-    let id = e.target.parentElement.getAttribute('data-id');
-    db.collection('trees').doc(id).delete();
-  })
-  newTree.update.addEventListener('click', (e) => {
-    e.stopPropagation();
-    let id = e.target.parentElement.getAttribute('data-id');
-    treeUpdate(newTree.content, id);
-  })
-  newTree.show.addEventListener('click', (e) => {
-    e.stopPropagation();
-    let id = e.target.parentElement.getAttribute('data-id');
-    showBranches(newTree.content, id)
-  })
-
-  treeList.appendChild(newTree.content);
+  if(doc){
+    const newTree = new Tree(
+      doc.data().title,
+      doc.data().description,
+      doc.data().color
+    )
+  
+    newTree.content.setAttribute('data-id', doc.id);
+  
+    newTree.remove.addEventListener('click', (e) => {
+      e.stopPropagation();
+      let id = e.target.parentElement.getAttribute('data-id');
+      db.collection('trees').doc(id).delete();
+    })
+    newTree.update.addEventListener('click', (e) => {
+      e.stopPropagation();
+      let id = e.target.parentElement.getAttribute('data-id');
+      treeUpdate(newTree.content, id);
+    })
+    newTree.show.addEventListener('click', (e) => {
+      e.stopPropagation();
+      let id = e.target.parentElement.getAttribute('data-id');
+      showBranches(newTree.content, id)
+    })
+  
+    treeList.appendChild(newTree.content);
+  }else {
+    let treeMessage = document.createElement('h2');
+    treeMessage.textContent = 'No Projects, create one from the navbar.'
+    treeList.appendChild(treeMessage);
+  }
 }
 
 const formEvents = (user) => {
@@ -148,7 +197,8 @@ const formEvents = (user) => {
       status: false,
       due_date: formBranch.form.due_date.value,
       notes: formBranch.form.notes.value,
-      tree_id: id
+      tree_id: id,
+      user_id: user.uid
     }).then(() => {
       formBranch.form.reset();
     }).catch(err => {
@@ -169,9 +219,16 @@ const formEvents = (user) => {
       console.log(err.message);
     })
   })
+
+  viewAll.addEventListener('click', (e) => {
+    e.stopPropagation()
+    addAll(user);
+  })
+
+  addAll(user);
 }
 
-treeWrapper.appendChild(formTree.content);
+//treeWrapper.appendChild(formTree.content);
 treeWrapper.appendChild(treeList);
 
-export  {renderTree, formEvents, branchWrapper, treeWrapper};
+export  {renderTree, formEvents, branchWrapper, treeWrapper, formTree};
